+++ cpp/FMU-proxy/example/CMakeLists.txt
@@ -3,25 +3,25 @@ cmake_minimum_required(VERSION 3.10)
 include_directories(${FMU_PROXY_SOURCE_DIR}/include)
 
 add_executable(Benchmark
-        Benchmark.cpp
-        TestUtil.cpp)
+        benchmark.cpp
+        test_util.cpp)
 target_link_libraries(Benchmark fmi)
 
 add_executable(FmuTest
-        FmuTest.cpp
-        TestUtil.cpp)
+        fmu_test.cpp
+        test_util.cpp)
 target_link_libraries(FmuTest fmi)
 
 add_executable(HeartBeatTest
-        HeartBeatTest.cpp
-        TestUtil.cpp)
+        heartbeat_test.cpp
+        test_util.cpp)
 target_link_libraries(HeartBeatTest heartbeat fmi pthread)
 
 add_executable(ThriftClientTest
-        ThriftClientTest.cpp)
+        thrift_client_test.cpp)
 target_link_libraries(ThriftClientTest thrift-client)
 
 
 add_executable(ThriftServerTest
-        ThriftServerTest.cpp)
+        thrift_server_test.cpp)
 target_link_libraries(ThriftServerTest thrift-server pthread)
