+++ cpp/FMU-proxy/src/fmuproxy/fmi/fmu_helper.cpp
