+++ cpp/FMU-proxy/src/fmuproxy/heartbeat/heartbeat_helper.cpp
@@ -1,0 +1,90 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2017-2018 Norwegian University of Technology
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING  FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+#include <string>
+#include <sstream>
+#include <iomanip>
+#include <boost/uuid/uuid.hpp>
+#include <boost/uuid/uuid_generators.hpp>
+#include <boost/uuid/uuid_io.hpp>
+
+using namespace std;
+using namespace boost::uuids;
+
+namespace fmuproxy::heartbeat {
+
+    const string generate_uuid() {
+        random_generator generator;
+        uuid uuid = generator();
+        return to_string(uuid);
+    }
+
+    //https://stackoverflow.com/questions/7724448/simple-json-string-escape-for-c
+    string escape_json(const std::string &s) {
+        std::ostringstream o;
+        for (auto c = s.cbegin(); c != s.cend(); c++) {
+            if (*c == '"' || *c == '\\' || ('\x00' <= *c && *c <= '\x1f')) {
+                o << "\\u"
+                  << std::hex << std::setw(4) << std::setfill('0') << (int)*c;
+            } else {
+                o << *c;
+            }
+        }
+        return o.str();
+    }
+
+    // https://stackoverflow.com/questions/44994203/how-to-get-the-http-response-string-using-curl-in-c
+    size_t write_callback(char *contents, size_t size, size_t nmemb, void *userp)
+    {
+        ((std::string*)userp)->append((char*)contents, size * nmemb);
+        return size * nmemb;
+    }
+
+
+    //https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
+    // trim from start (in place)
+    static inline void ltrim(string &s) {
+        s.erase(s.begin(), std::find_if(s.begin(), s.end(), [](int ch) {
+            return !std::isspace(ch);
+        }));
+    }
+
+    //https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
+    // trim from end (in place)
+    static inline void rtrim(string &s) {
+        s.erase(std::find_if(s.rbegin(), s.rend(), [](int ch) {
+            return !std::isspace(ch);
+        }).base(), s.end());
+    }
+
+    //https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
+    // trim from both ends (in place)
+    static inline void trim(string &s) {
+        ltrim(s);
+        rtrim(s);
+    }
+
+
+
+}
