+++ cpp/FMU-proxy/src/fmuproxy/thrift/client/ThriftClient.cpp
@@ -32,7 +32,7 @@
 #include <fmuproxy/thrift/common/definitions_types.h>
 #include <fmuproxy/thrift/client/ThriftClient.hpp>
 
-#include "ThriftClientHelper.cpp"
+#include "thrift_client_helper.cpp"
 
 using namespace std;
 using namespace fmuproxy::thrift::client;
