+++ cpp/FMU-proxy/src/fmuproxy/thrift/client/thrift_client_helper.cpp
