+++ cpp/FMU-proxy/src/fmuproxy/thrift/server/thrift_server_helper.cpp
@@ -24,8 +24,6 @@
 
 #include <cfloat>
 #include <fmilib.h>
-
-#include <fmuproxy/fmi/fmi_definitions.hpp>
 #include <fmuproxy/fmi/fmi_definitions.hpp>
 #include <fmuproxy/thrift/common/definitions_types.h>
